Better handle ADC settings in unusual configurations.
This PR extends the coverage of the ray-traced tan2fp ZB fits to handle angle differences up to 360 degrees.  These are seldom used in practice but are possible, and produce different distortion patterns.  The PR is also slightly more careful about how ADC angles are translated into look ups into the fits.

Additionally, the ray tracing code was parallelized to speed it up.